user_manager/requirements.txt
flake8==3.9.2
flask==1.1.2
flask_sqlalchemy==2.4.4
flask_marshmallow==0.14.0
marshmallow_sqlalchemy==0.24.1
psycopg2-binary==2.8.6
flask_migrate==2.5.3
pyjwt==2.0.1
flask-bcrypt==0.7.1
boto3==1.16.61
flask_seeder
requests
flask_cors
pytest
coverage
pandas
pycryptodome==3.9.8
pandas==1.2.0
numpy==1.17.4
pytz==2021.3
python-dotenv
Flask-Caching==1.10.1
SQLAlchemy==1.4.18
Flask-APScheduler==1.12.2
docxtpl==0.5.15
python-docx==0.8.6
